Type of raw values should be "long"
git-svn-id: https://svn.52north.org/svn/ilwis/drawers@1847 b64c5c0c-1e74-440b-9268-4ad09cf26390

+++ Sources/SpaceTimeCube/SpaceTimeDrawer.h
@@ -42,7 +42,7 @@ class _export SpaceTimeDrawer : public FeatureLayerDrawer, public TemporalDrawer
 		BaseMapPtr *basemap;
 		bool fValueMap;
 		CoordBounds cube;
-		vector<int> disabledRaws;
+		vector<long> disabledRaws;
 		bool *fHatching;
 	private:
 		Column prevAttColumn;
